% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_animator.R
\name{play_animator}
\alias{play_animator}
\title{Play Animator}
\usage{
play_animator(ds, game_id, play_id, play_details, color_scheme)
}
\arguments{
\item{ds}{arrow dataset}

\item{game_id}{integer}

\item{play_id}{integer}

\item{play_details}{data.frame}

\item{color_scheme}{data.frame}
}
\description{
Play Animator
}
