% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_viz.R
\name{zone_builder}
\alias{zone_builder}
\title{Zone Builder}
\usage{
zone_builder(x_breaks, y_breaks)
}
\arguments{
\item{x_breaks}{numeric}

\item{y_breaks}{numeric}
}
\value{
zone_grid
}
\description{
Zone Builder
}
