% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_viz.R
\name{zone_tile}
\alias{zone_tile}
\title{Zone Tiles}
\usage{
zone_tile(x_breaks, y_breaks, zone_grid, plot_width_limits, plot_height_limits)
}
\arguments{
\item{x_breaks}{numeric}

\item{y_breaks}{numeric}

\item{zone_grid}{data.frame}

\item{plot_width_limits}{numeric}

\item{plot_height_limits}{numeric}
}
\value{
zone_tiles_plot
}
\description{
Zone Tiles
}
